requirements.txt
pandas>=2.0.0
python-dotenv>=1.0.0
requests-oauthlib>=1.3.1
openpyxl>=3.1.5  # Required for pandas Excel support